perf(deepseek_v4): drop fp32 casts in Indexer score computation
The Indexer's score matmul cast q and pooled to fp32, then the per-head
weight scaling promoted weights to fp32 too. This forces an fp32 GEMM
(double the bandwidth and FLOPs vs bf16) on a hot decode-time op whose
cost scales with pooled-cache size — measured 6.6% of decode at 32K and
extrapolating to ~19% at 100K context on DeepSeek-V4-Flash-6bit.

Drop the explicit casts. MLX's bf16 GEMM already accumulates in fp32
internally, so the matmul's numerical behavior is preserved up to the
final downcast. The downstream argpartition top-k selection is robust to
small score perturbations — top-k routing decisions are inherently fuzzy
at the boundary, so bf16 precision is sufficient.

+++ mlx_lm/models/deepseek_v4.py
@@ -1606,11 +1606,9 @@ class Indexer(nn.Module):
         q = q.transpose(0, 2, 1, 3)
         q = _apply_partial_rope(q, position_rope, offset)
 
-        scores = q.astype(mx.float32) @ pooled[:, None].swapaxes(-1, -2).astype(
-            mx.float32
-        )
+        scores = q @ pooled[:, None].swapaxes(-1, -2)
         scores = mx.maximum(scores, 0) * self.scale
-        weights = self.weights_proj(x).astype(mx.float32) * (self.n_heads**-0.5)
+        weights = self.weights_proj(x) * (self.n_heads**-0.5)
         scores = (scores * weights.swapaxes(-1, -2)[..., None]).sum(axis=1)
         lengths = cache.pooled_lengths("indexer_state") if cache is not None else None
         if lengths is not None:
